Update Meal plan.py
Added in a check to see if an ingredient had already been used and adds it if it hasnt. Also returns the Unique ID identifier if it hasnt so i can add it it the recipie table. Thsi will let me more easily in future connect different recipies and they will all use the same number system.