Update Meal plan.py
It now works to add recipies to a table. Using tabular i have nice looking tables to display it too which is nice. Added view recipies def which is one of the options on the menu i want to have. Going to add a view all tables, delete tables, edit tables next. Also working with connecting the ingredients tables with the ID numbers finally.