Update Meal plan.py
Added in a few bits and bobs, basically got the recipie part down now. Few bits on error handeling need to be added but apart from that all good there. Working on the meal plan part now. Got it to add a meal plan table with the days as primary keys so they wont be reperated. You can edit the meals on each day and save. Next up is having a table of all the Recipies that have been added and cross checkign them in the week so i can create an itemised shoppig list