Update Meal plan.py
IT FUCKING WORKSSSSSSS. The portion part was easier than i though and is now working! Next step is to add some units if i want to and go through and add in more error handling such as if the input is blank or the wrong type. I had an idea that i create or find a data base with up to date ingredient prices (maybe web scraping)  and find a cost for the shopping list.